move localstorage utils
